+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Bump k8scc to 15.0.0.
 - Disable kube-proxy on release v19 and newer.
 
 ## [13.2.1] - 2022-08-31